Refactor identifiers.go into category-specific files
Split the 1,863-line monolithic identifiers.go into 10 files organized
by test category: accesscontrol, certification, lifecycle, manageability,
networking, observability, operator, performance, and platform.

The shared identifiers.go now contains only constants, helper functions,
and package-level maps (~73 lines). Each category file declares its own
identifiers and registers them via init().

All 102 AddCatalogEntry registrations are preserved. Identifiers within
each file are sorted alphabetically. No import changes required in
consuming packages since all files remain in the same package.
